chore.更新企鹅识别库到 v4.2.3 版本，修复磨难关卡的识别

+++ 3rdparty/include/penguin-stats-recognize/recognize.hpp
@@ -15,7 +15,6 @@ using dict = nlohmann::ordered_json;
 
 namespace penguin
 {
-
 enum class StatusFlags
 {
     NORMAL = 0,
@@ -823,8 +822,8 @@ private:
         topleft_new.x *= coeff_multiinv;
         topleft_new.y *= coeff_multiinv;
         cv::Size size_new = cv::Size(
-            round(TEMPLATE_WIDTH * ((double)_diameter / TEMPLATE_DIAMETER)),
-            round(TEMPLATE_HEIGHT * ((double)_diameter / TEMPLATE_DIAMETER)));
+            static_cast<int>(round(TEMPLATE_WIDTH * ((double)_diameter / TEMPLATE_DIAMETER))),
+            static_cast<int>(round(TEMPLATE_HEIGHT * ((double)_diameter / TEMPLATE_DIAMETER))));
         if (topleft_new.x + size_new.width > width)
         {
             size_new.width = width - topleft_new.x;
@@ -843,9 +842,9 @@ private:
     }
     void _get_quantity()
     {
-        cv::Rect quantityrect = cv::Rect(0, round(height * _ITEM_QTY_Y_PROP),
-                                         round(width * _ITEM_QTY_WIDTH_PROP),
-                                         round(height * _ITEM_QTY_HEIGHT_PROP));
+        cv::Rect quantityrect = cv::Rect(0, static_cast<int>(round(height * _ITEM_QTY_Y_PROP)),
+                                         static_cast<int>(round(width * _ITEM_QTY_WIDTH_PROP)),
+                                         static_cast<int>(round(height * _ITEM_QTY_HEIGHT_PROP)));
         cv::Mat quantityimg = _img(quantityrect);
         _quantity.set_img(quantityimg);
         _quantity.analyze();
